Renamed glue2/types.py so that in python 3.9 and greater, it doesn't conflict with functool's types.py.  (Also changed IPF references from types to glue2types)
