+++ ipf/glue2/valet.py
@@ -24,7 +24,7 @@ import hashlib
 from ipf.error import StepError
 from ipf.urnprefix import IPF_URN_PREFIX
 from . import application
-from .types import AppEnvState, ApplicationHandle
+from .glue2types import AppEnvState, ApplicationHandle
 
 #######################################################################################################################
